cp-427: * move check for id path to service for chat and journal

+++ backend/src/packages/chats/chat.service.ts
@@ -168,6 +168,13 @@ class ChatService implements Service {
     id: number;
     userId: number;
   }): Promise<boolean> {
+    if (!Number(id)) {
+      throw new ChatError({
+        status: HTTPCode.BAD_REQUEST,
+        message: ExceptionMessage.CHAT_NOT_FOUND,
+      });
+    }
+
     const deletedCount = await this.chatRepository.delete({ id, userId });
     if (!deletedCount) {
       throw new ChatError({
